Unbreak dashboard Next.js build
The dashboard build has been broken on main since ~2026-04-16 because
several route files exported arbitrary symbols (forbidden by App
Router), had a broken relative import, and called query() with a second
parameter the function never accepted. None of this was caught by CI
because CI does not run `next build`.

Fixes:
- Move computeAnomaly out of app/api/anomaly-check/route.ts into
  lib/anomaly.ts; route and test import from there.
- Move parseSeason/Season out of app/api/seasons/route.ts into
  lib/seasons.ts; route and test import from there.
- Correct the EtlSyncRun import path in
  app/api/etl/runs/[id]/route.ts (../types → ../../types).
- Extend query() in lib/db.ts to accept optional parameterised values
  so the ETL runs/stats routes type-check against their pg-style $1,
  $2 call sites.
- Normalise the x-axis key in mergeComparisonSeries to a string so
  the return shape matches its declared type.

Verified end-to-end: docker compose build dashboard succeeds, container
recreates, HTTP 200 on /, and the Monitor ETL link (from #219) is now
visible in the rendered sidebar.
